requirements.txt
python-dotenv~=1.2.1
aiogram~=3.22.0
requests~=2.32.5